Fix issues with missing arguments
In DAP protocol, the arguments key is optional. With dape, I run
into issues when in this case, the DebugAdapter has failed with:

```
ERROR: KeyError: key "arguments" not found
Stacktrace:
 [1] getindex
   @ ./dict.jl:498 [inlined]
 [2] dispatch_msg(x::DebugAdapter.DAPRPC.DAPEndpoint{Sockets.TCPSocket, Sockets.TCPSocket}, dispatcher::DebugAdapter.DAPRPC.MsgDispatcher, msg::Dict{String, Any})
   @ DebugAdapter.DAPRPC ~/active/projects/julia/DebugAdapter.jl/src/DAPRPC/typed.jl:86
 [3] (::DebugAdapter.var"#132#161"{Dict{String, Any}, Nothing, DebugAdapter.DAPRPC.MsgDispatcher, DebugAdapter.DAPRPC.DAPEndpoint{Sockets.TCPSocket, Sockets.TCPSocket}})()
   @ DebugAdapter ~/active/projects/julia/DebugAdapter.jl/src/packagedef.jl:134
```

Also, for optional fields, the dape client sometimes sends null values instead of omitting the key.

Making the code a big more defensive sounds like good idea and if fixes
the issue for me.