Fix world age errors with Enums
If code defines methods which are needed for printing the value in the debug view, we seem to have a world age problem. This PR addresses the issue by always evaluating the value in the latest world age.

There probably is a more elegant or at least a more general/systematic solution to the problem. However, as the debugger cannot be uses at all in code which uses Enums for more than two years, I think it's important to fix this quickly.